ai_nutricare/requirements.txt
streamlit>=1.35.0
groq>=0.9.0
pypdfium2>=4.30.0
Pillow>=10.0.0
python-dotenv>=1.0.0
